# optimize upsample bicubic mode performance on CPU