This PR aims at improving upsample bicubic mode performance on CPU. This patch simply parallel the outer loop on CPU, but sitll have performance improvement for single thread case since the original code has non-contigous memory access pattern which is not efficient. On my test machine single thread and single socket achieve 7x and 117x speedup respectively.
Related work: https://github.com/pytorch/pytorch/pull/31452, https://github.com/pytorch/pytorch/pull/34864

Result on Intel(R) Xeon(R) Gold 6248 CPU @ 2.50GHz, 20 cores per socket, dual socket.

```bash
before:
### using OMP_NUM_THREADS=20
### using numactl --physcpubind=0-19 --membind=0

input.is_contiguous(memory_format=torch.channels_last):  False
input.is_contiguous():  True
forward time per iteration: 10105.570 ms
upsample_bilinear_cl: memory format: nhwc, input size:  torch.Size([32, 128, 64, 64])
input.is_contiguous(memory_format=torch.channels_last):  True
input.is_contiguous():  False
forward time per iteration: 10092.228 ms

### using OMP_NUM_THREADS=1
### using numactl --physcpubind=0 --membind=0

upsample_bilinear: memory format: nchw, input size:  torch.Size([32, 128, 64, 64])
input.is_contiguous(memory_format=torch.channels_last):  False
input.is_contiguous():  True
forward time per iteration: 10167.732 ms
upsample_bilinear_cl: memory format: nhwc, input size:  torch.Size([32, 128, 64, 64])
input.is_contiguous(memory_format=torch.channels_last):  True
input.is_contiguous():  False
forward time per iteration: 10199.118 ms


after:
### using OMP_NUM_THREADS=20
### using numactl --physcpubind=0-19 --membind=0

upsample_bilinear: memory format: nchw, input size:  torch.Size([32, 128, 64, 64])
input.is_contiguous(memory_format=torch.channels_last):  False
input.is_contiguous():  True
forward time per iteration: 86.109 ms
upsample_bilinear_cl: memory format: nhwc, input size:  torch.Size([32, 128, 64, 64])
input.is_contiguous(memory_format=torch.channels_last):  True
input.is_contiguous():  False
forward time per iteration: 139.421 ms

### using OMP_NUM_THREADS=1
### using numactl --physcpubind=0 --membind=0

upsample_bilinear: memory format: nchw, input size:  torch.Size([32, 128, 64, 64])
input.is_contiguous(memory_format=torch.channels_last):  False
input.is_contiguous():  True
forward time per iteration: 1449.662 ms
upsample_bilinear_cl: memory format: nhwc, input size:  torch.Size([32, 128, 64, 64])
input.is_contiguous(memory_format=torch.channels_last):  True
input.is_contiguous():  False
forward time per iteration: 2072.506 ms
```